Improved structure
Moved profanity check to the top.
Added username to pong to differentiate if the bot is running twice.
Wrapped all the message processing code in a try:except block and log errors to the bot log channel. Before errors were logged to the device running the bot.
Added comments.
Rearranged imports.