Improve the spam detection (#7)
If a user sends a message to three channels in a 5 minute window containing a link or attachment they will be banned.
These requirements are a bit loose but we can deal with any false positives manually. I don't think there will be many, if any.
The latest spam examples only have image attachments and the images are slightly different so there is no easy way to detect this.